# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Programação de Entrega
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Delivery unit
 */
export type EntregaUnidade = typeof EntregaUnidade[keyof typeof EntregaUnidade];


export const EntregaUnidade = {
  MATRIZ: 'MATRIZ',
  FILIAL: 'FILIAL',
  'MATRIZ_+_FILIAL': 'MATRIZ + FILIAL',
} as const;

/**
 * V column flag — null=empty, V=concluido, 2A=segunda entrega
 * @nullable
 */
export type EntregaV = typeof EntregaV[keyof typeof EntregaV] | null;


export const EntregaV = {
  V: 'V',
  '2A': '2A',
} as const;

/**
 * S column: none=filled (black), confirmed (tick)
 */
export type EntregaChecked = typeof EntregaChecked[keyof typeof EntregaChecked];


export const EntregaChecked = {
  NONE: 'none',
  FILLED: 'filled',
  CONFIRMED: 'confirmed',
} as const;

export interface Entrega {
  id: number;
  /** Date in YYYY-MM-DD format */
  date: string;
  sortOrder: number;
  /** S column: none, filled (black), confirmed (tick) */
  checked: EntregaChecked;
  cliente: string;
  /**
     * Hour annotation (e.g. 7H, 11MT) — highlighted yellow when filled
     * @nullable
     */
  hrs: string | null;
  /**
     * Observations (e.g. 001 + 002, 087)
     * @nullable
     */
  obs: string | null;
  /**
     * Driver name
     * @nullable
     */
  motorista: string | null;
  /**
     * Vehicle license plate
     * @nullable
     */
  placa: string | null;
  /** Delivery unit */
  unidade: EntregaUnidade;
  /** NF column: none, x (red), check (green) */
  nf: "none" | "x" | "check";
  /** CG column: none, x (red), check (green) */
  cg: "none" | "x" | "check";
  /**
     * V column flag — null=empty, V=concluido, 2A=segunda entrega
     * @nullable
     */
  v: EntregaV;
  /**
     * Divergencias notes
     * @nullable
     */
  divergencias: string | null;
}

export type EntregaInputUnidade = typeof EntregaInputUnidade[keyof typeof EntregaInputUnidade];


export const EntregaInputUnidade = {
  MATRIZ: 'MATRIZ',
  FILIAL: 'FILIAL',
  'MATRIZ_+_FILIAL': 'MATRIZ + FILIAL',
} as const;

/**
 * @nullable
 */
export type EntregaInputV = typeof EntregaInputV[keyof typeof EntregaInputV] | null;


export const EntregaInputV = {
  V: 'V',
  '2A': '2A',
} as const;

export interface EntregaInput {
  date: string;
  /** @nullable */
  sortOrder?: number | null;
  checked?: EntregaChecked;
  cliente: string;
  /** @nullable */
  hrs?: string | null;
  /** @nullable */
  obs?: string | null;
  /** @nullable */
  motorista?: string | null;
  /** @nullable */
  placa?: string | null;
  unidade: EntregaInputUnidade;
  nf?: "none" | "x" | "check";
  cg?: "none" | "x" | "check";
  /** @nullable */
  v?: EntregaInputV;
  /** @nullable */
  divergencias?: string | null;
}

/**
 * @nullable
 */
export type EntregaUpdateUnidade = typeof EntregaUpdateUnidade[keyof typeof EntregaUpdateUnidade] | null;


export const EntregaUpdateUnidade = {
  MATRIZ: 'MATRIZ',
  FILIAL: 'FILIAL',
  'MATRIZ_+_FILIAL': 'MATRIZ + FILIAL',
} as const;

/**
 * @nullable
 */
export type EntregaUpdateV = typeof EntregaUpdateV[keyof typeof EntregaUpdateV] | null;


export const EntregaUpdateV = {
  V: 'V',
  '2A': '2A',
} as const;

export interface EntregaUpdate {
  checked?: EntregaChecked;
  /** @nullable */
  cliente?: string | null;
  /** @nullable */
  hrs?: string | null;
  /** @nullable */
  obs?: string | null;
  /** @nullable */
  motorista?: string | null;
  /** @nullable */
  placa?: string | null;
  /** @nullable */
  unidade?: EntregaUpdateUnidade;
  nf?: "none" | "x" | "check";
  cg?: "none" | "x" | "check";
  /** @nullable */
  v?: EntregaUpdateV;
  /** @nullable */
  sortOrder?: number | null;
  /** @nullable */
  divergencias?: string | null;
}

export interface Motorista {
  id: number;
  nome: string;
  placa: string;
}

export interface MotoristasInput {
  nome: string;
  placa: string;
}

export interface ReorderInput {
  /** Ordered list of entrega IDs representing the new sort order */
  ids: number[];
}

export type ListEntregasParams = {
/**
 * Date in YYYY-MM-DD format (defaults to today)
 */
date?: string;
};
