# lib/api-zod/src/generated/types/entrega.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Programação de Entrega
 * OpenAPI spec version: 0.1.0
 */
import type { EntregaUnidade } from './entregaUnidade';
import type { EntregaV } from './entregaV';
import type { EntregaChecked } from './entregaChecked';

export interface Entrega {
  id: number;
  /** Date in YYYY-MM-DD format */
  date: string;
  sortOrder: number;
  /** S column: none, filled (black), confirmed (tick) */
  checked: EntregaChecked;
  cliente: string;
  /**
     * Hour annotation (e.g. 7H, 11MT) — highlighted yellow when filled
     * @nullable
     */
  hrs: string | null;
  /**
     * Observations (e.g. 001 + 002, 087)
     * @nullable
     */
  obs: string | null;
  /**
     * Driver name
     * @nullable
     */
  motorista: string | null;
  /**
     * Vehicle license plate
     * @nullable
     */
  placa: string | null;
  /** Delivery unit */
  unidade: EntregaUnidade;
  /** NF column: none, x (red), check (green) */
  nf: "none" | "x" | "check";
  /** CG column: none, x (red), check (green) */
  cg: "none" | "x" | "check";
  /**
     * V column flag — null=empty, V=concluido, 2A=segunda entrega
     * @nullable
     */
  v: EntregaV;
  /**
     * Divergencias notes
     * @nullable
     */
  divergencias: string | null;
}