# lib/api-zod/src/generated/types/entregaInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Programação de Entrega
 * OpenAPI spec version: 0.1.0
 */
import type { EntregaInputUnidade } from './entregaInputUnidade';
import type { EntregaInputV } from './entregaInputV';
import type { EntregaChecked } from './entregaChecked';

export interface EntregaInput {
  date: string;
  /** @nullable */
  sortOrder?: number | null;
  checked?: EntregaChecked;
  cliente: string;
  /** @nullable */
  hrs?: string | null;
  /** @nullable */
  obs?: string | null;
  /** @nullable */
  motorista?: string | null;
  /** @nullable */
  placa?: string | null;
  unidade: EntregaInputUnidade;
  nf?: "none" | "x" | "check";
  cg?: "none" | "x" | "check";
  /** @nullable */
  v?: EntregaInputV;
  /** @nullable */
  divergencias?: string | null;
}