# lib/api-zod/src/generated/types/entregaUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Programação de Entrega
 * OpenAPI spec version: 0.1.0
 */
import type { EntregaUpdateUnidade } from './entregaUpdateUnidade';
import type { EntregaUpdateV } from './entregaUpdateV';
import type { EntregaChecked } from './entregaChecked';

export interface EntregaUpdate {
  checked?: EntregaChecked;
  /** @nullable */
  cliente?: string | null;
  /** @nullable */
  hrs?: string | null;
  /** @nullable */
  obs?: string | null;
  /** @nullable */
  motorista?: string | null;
  /** @nullable */
  placa?: string | null;
  /** @nullable */
  unidade?: EntregaUpdateUnidade;
  nf?: "none" | "x" | "check";
  cg?: "none" | "x" | "check";
  /** @nullable */
  v?: EntregaUpdateV;
  /** @nullable */
  sortOrder?: number | null;
  /** @nullable */
  divergencias?: string | null;
}