# lib/api-zod/src/generated/types/motoristasInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Programação de Entrega
 * OpenAPI spec version: 0.1.0
 */

export interface MotoristasInput {
  nome: string;
  placa: string;
}